// client-fe/src/app/model/class/discount.class.ts
export class Discount {
  id: number;
  maPhieuGiamGia: string;
  tenPhieuGiamGia: string;
  kieu: number;
  loai: number;
  giaTri: number;
  giaTriMax: number;
  dieuKienGiam: number;
  soLuong: number;
  thoiGianBatDau: string;
  thoiGianKetThuc: string;
  trangThai: string;
  isSelected: boolean;
}